Added "materialize" CLI argument and API function.
The materialize function is basically a bag bootstrapper. When invoked, it will attempt to fully materialize a bag by performing multiple actions depending on the context of the input <path>. If <path> is a URL or a URI of a resolvable identifier scheme, the file referenced by this value will first be downloaded to the current directory. Next, if the <path> value (or previously downloaded file) is a local path to a supported archive format, the archive will be extracted to the current directory. Then, if the <path> value (or previously extracted file) is a valid bag directory, any remote file references contained within the bag's "fetch.txt" file will attempt to be resolved. Finally, full validation will be run on the materialized bag. If any one of these steps fail, an error is raised.

Change return value of fetch_file/get_file API to be the path of the successfully retrieved file, or None, instead of boolean.
Changed BaseResolver to not automatically set "Accept: application/json" and updated individual resolvers to set it accordingly when necessary.
Updated/added unit tests.